Start updating RFC to match changes
